move all tests to src/test/java and mark junit as test dependency
