--- nd4j-api/src/main/java/org/nd4j/linalg/indexing/conditions/test/ConditionTest.java
@@ -1,46 +1,0 @@
-/*
- * Copyright 2015 Skymind,Inc.
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package org.nd4j.linalg.indexing.conditions.test;
-
-import org.junit.Test;
-import org.nd4j.linalg.api.ndarray.INDArray;
-import org.nd4j.linalg.factory.Nd4j;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * Created by agibsonccc on 10/10/14.
- */
-public abstract class ConditionTest {
-
-    @Test
-    public void testNeq() {
-        INDArray n = Nd4j.create(new float[]{1, 2, 3, 4});
-        INDArray n2 = n.neq(Nd4j.create(new int[]{4}));
-        assertEquals(4, n2.sum(Integer.MAX_VALUE).getFloat(0), 1e-1);
-
-    }
-
-    @Test
-    public void testEq() {
-        INDArray n = Nd4j.create(new float[]{1, 2, 3, 4});
-        INDArray n2 = n.eq(Nd4j.create(new int[]{4}));
-        assertEquals(0, n2.sum(Integer.MAX_VALUE).getFloat(0), 1e-1);
-
-    }
-
-}
