+++ nd4j-api/src/test/java/org/nd4j/linalg/api/NDArrayTests.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.nd4j.linalg.api.test;
+package org.nd4j.linalg.api;
 
 
 import org.junit.After;
@@ -1255,7 +1255,7 @@ public abstract class NDArrayTests {
     @Test
     public void testRand() {
         INDArray rand = Nd4j.randn(5, 5);
-        Nd4j.getDistributions().createUniform(0.4, -0.4).sample(5);
+        Nd4j.getDistributions().createUniform(-0.4, 0.4).sample(5);
         Nd4j.getDistributions().createNormal(1, 5).sample(10);
         Nd4j.getDistributions().createBinomial(5, 1.0).sample(new int[]{5, 5});
         Nd4j.getDistributions().createBinomial(1, Nd4j.ones(5, 5)).sample(rand.shape());
