+++ nd4j-api/src/test/java/org/nd4j/linalg/fft/BaseIFFTTests.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.nd4j.linalg.fft.test;
+package org.nd4j.linalg.fft;
 
 import org.junit.Test;
 import org.nd4j.linalg.api.complex.IComplexNDArray;
