+++ nd4j-api/src/test/java/org/nd4j/linalg/learning/AdaGradTest.java
