+++ nd4j-api/src/test/java/org/nd4j/linalg/lossfunctions/LossFunctionTests.java
@@ -16,13 +16,14 @@
 
 package org.nd4j.linalg.lossfunctions;
 
+import static org.junit.Assert.assertEquals;
+
 import org.junit.Test;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.factory.Nd4j;
+import org.nd4j.linalg.lossfunctions.LossFunctions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import static org.junit.Assert.assertEquals;
 
 /**
  * Testing loss function
