+++ nd4j-jblas/src/test/java/org/nd4j/linalg/jblas/NDArrayTests.java
@@ -34,7 +34,7 @@ import static org.junit.Assert.assertEquals;
  *
  * @author Adam Gibson
  */
-public class NDArrayTests extends org.nd4j.linalg.api.test.NDArrayTests {
+public class NDArrayTests extends org.nd4j.linalg.api.NDArrayTests {
     private static Logger log = LoggerFactory.getLogger(NDArrayTests.class);
