+++ nd4j-jblas/src/test/java/org/nd4j/linalg/jblas/convolution/ConvolutionTests.java
@@ -19,5 +19,5 @@ package org.nd4j.linalg.jblas.convolution;
 /**
  * Created by agibsonccc on 9/6/14.
  */
-public class ConvolutionTests extends org.nd4j.linalg.convolution.test.ConvolutionTests {
+public class ConvolutionTests extends org.nd4j.linalg.convolution.ConvolutionTests {
 }
