+++ nd4j-jblas/src/test/java/org/nd4j/linalg/jblas/fft/test/FFTTests.java
@@ -16,7 +16,7 @@
 
 package org.nd4j.linalg.jblas.fft.test;
 
-import org.nd4j.linalg.fft.test.BaseFFTTest;
+import org.nd4j.linalg.fft.BaseFFTTest;
 
 /**
  * Created by agibsonccc on 9/5/14.
