+++ nd4j-jblas/src/test/java/org/nd4j/linalg/jblas/fft/test/IFFTTests.java
@@ -16,7 +16,7 @@
 
 package org.nd4j.linalg.jblas.fft.test;
 
-import org.nd4j.linalg.fft.test.BaseIFFTTests;
+import org.nd4j.linalg.fft.BaseIFFTTests;
 
 /**
  * Created by agibsonccc on 9/6/14.
