+++ nd4j-jblas/src/test/java/org/nd4j/linalg/jblas/indexing/ConditionTests.java
@@ -16,7 +16,7 @@
 
 package org.nd4j.linalg.jblas.indexing;
 
-import org.nd4j.linalg.indexing.conditions.test.ConditionTest;
+import org.nd4j.linalg.indexing.conditions.ConditionTest;
 
 /**
  * Created by agibsonccc on 10/10/14.
